move clustermanager package
