+++ testutils/clustermanager/e2e-tests/boskos/boskos_test.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 	"testing"
 
-	"knative.dev/pkg/testutils/common"
+	"knative.dev/pkg/testutils/clustermanager/e2e-tests/common"
 )
 
 var (
