+++ testutils/clustermanager/e2e-tests/client.go
