+++ testutils/clustermanager/e2e-tests/gke.go
@@ -24,8 +24,8 @@ import (
 
 	container "google.golang.org/api/container/v1beta1"
 	"knative.dev/pkg/test/gke"
-	"knative.dev/pkg/testutils/prow/clustermanager/boskos"
-	"knative.dev/pkg/testutils/common"
+	"knative.dev/pkg/testutils/clustermanager/e2e-tests/boskos"
+	"knative.dev/pkg/testutils/clustermanager/e2e-tests/common"
 )
 
 const (
