+++ testutils/clustermanager/e2e-tests/util_test.go
@@ -19,7 +19,7 @@ package clustermanager
 import (
 	"testing"
 
-	"knative.dev/pkg/testutils/common"
+	"knative.dev/pkg/testutils/clustermanager/e2e-tests/common"
 )
 
 func TestGetResourceName(t *testing.T) {
